Remove all the commented, non-SampleUtil code across the engine.

+++ mixxx/src/engine/enginevolume.cpp
@@ -35,23 +35,8 @@ EngineVolume::~EngineVolume()
 void EngineVolume::process(const CSAMPLE * pIn, const CSAMPLE * pOut, const int iBufferSize)
 {
     CSAMPLE * pOutput = (CSAMPLE *)pOut;
-    float volume=(float)potmeter->get();
+    double volume = potmeter->get();
 
     // SampleUtil handles aliased buffers and gains of 1 or 0.
     SampleUtil::copyWithGain(pOutput, pIn, volume, iBufferSize);
-
-    // if (volume == 1.)
-    // {
-    //     if (pIn!=pOut)
-    //     {
-    //         for (int i=0; i<iBufferSize; ++i)
-    //             pOutput[i] = pIn[i];
-    //         //memcpy(pOutput, pIn, sizeof(CSAMPLE) * iBufferSize);
-    //     }
-    // }
-    // else
-    // {
-    //     for (int i=0; i<iBufferSize; ++i)
-    //         pOutput[i] = pIn[i]*volume;
-    // }
 }
